1) Fixed compilation problems (only with certain compiler versions    it appears) for highscore manager. 2) Re-applied fix for warning in kart.hpp, which I had removed    by accident.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk@944 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/highscore_manager.cpp
@@ -131,7 +131,7 @@ void HighscoreManager::Save()
         lisp::Writer writer(filename);
         writer.beginList("highscores");
           writer.writeComment("Number of highscores in this file");
-          writer.write("number-entries\t",m_allScores.size());
+          writer.write("number-entries\t",(unsigned int)m_allScores.size());
           int record_number=0;
           for(type_all_scores::iterator i  = m_allScores.begin(); 
               i != m_allScores.end();  i++)
